Restructure indication for analysis types, such as probability, uncertainty, importance, and ccf.
The old implicit discovery of probability settings is abandoned.
Explicit settings are required, and only fault tree analysis to find
minimal cut sets is assumed to be the default request. This
restructuring breaks tests that assume that probability values are
calculated implicitly.

In addition, reporting of some analyses like importance are not yet
formatted in output.