Clarify FaultTree containers and functions
It is made clear that container keys are lower-case names of events,
parameters, groups, or components. This is in contrast to IDs used
in the main model container. This difference is made because FaultTree
and Component are used for scope-sensitive referencing and validation,
while the main model container has global reference and validation
for its members, which requires unique global ID.