Support event references with dots
References for events can include names of containers. This referencing
is resolved according to the rules of the OpenPSA MEF. The referencing
is scope-aware and role-aware. This is the expected behavior with
public and private roles of events for issue #107. However,
dereferencing procedures add performance overhead as noted in the
speedlog.