Refactor and cleanup the fault tree generator
Finding candidate common gates are turned into a generator function for
efficiency.

A reporting bug is corrected. The fraction of common gates and basic
events were calculated in a wrong way, using the total number of gates.
The right way is to use the relevant gates that have child gates or
basic events, respectively.

The code is cleaned with PyLint, and the documentation is updated.