Refactor the constant propagation functions
The constant propagation step does not rely on fault tree normalization
anymore. This change makes the logic of the constant propagation much
more complex due to the bigger variety of gate types in non-normalized
fault trees; however, the benefit of decoupling the normalization and
the constant propagation steps is that fault-tree optimization methods
are more useful with non-normalized fault trees. Since the constant
propagation is one of the first preprocessing techniques, its
requirement for the normalization makes the job of the optimization
techniques harder and less helpful for finding minimal cut sets.

The fault-tree optimization has more value than normalization,
house-event propagation, and other required steps. Therefore, the needs
of the optimization steps are of higher priority.

This implementation of the house-event propagation breaks the
ThreeMotor benchmark test due to assertion errors. The test is
disabled.