Refactor JoinGates function to be more flexible
The function does not require normalized, positive-gate-only,
preprocessed fault trees. It works with gates if they fit preprocessing
criteria. This means that the gate coalescing can be applied at any
stage of the overall fault tree preprocessing.

Modules are omitted from coalescing to preserve the positive work of the
modularization preprocessing steps.