Remove NOT gate logic from  PropagateComplements
The separation of NOT gate logic from complement propagation makes the
overall logic simpler and less error-prone. The NOT gate is handled by
the normalization functions to turn it into NULL type gate with a
negative sign.

This separation of concerns fixes #108. The top event is handled in a
special way for corner cases.