Transfer the publishing files from gh-pages
This transformation is done because there is no need to version control
the auto-generated files on gh-pages. All the scripts that must be
version controlled are transfered to the source branch of the website.

gh-pages branch must have only one commit with the recently generated
documentation.

This separation also saves space.