Introduce self-indexing IndexedFaultTree
This self-indexing feature helps decouple IndexedFaultTree from
FaultTreeAnalysis. The FaultTreeAnalysis should be agnostic about what
indices the indexed fault tree uses for its internal purposes.

The only contract is that indices for basic events start from 1 and end
with the number of basic events in the indexed fault tree. This helps
optimize preprocessing and analysis algorithms.