Use the auto-indexation in FaultTreeAnalysis
The indexation is delegated to the indexed fault tree with a contract
that basic events will start from 1. Replacing basic events with CCF
events is also done by the indexed fault tree.

Preprocessing tests and grouping modular children are disabled because
of the heavy reliance on the previous indexation logic. These facilities
will be recovered after refactoring of the code.