Use a struct for late defined input elements
Gates, parameters, basic events, and CCF groups are defined late due to
the requirement of their registered dependencies. Instead of one vector
to track these to-be-defined elements, a struct with multiple vectors
for different types of the late-defined elements avoids RTTI of the
elements.

This change is a part of the effort to minimize code with RTTI.