Implement Boolean Optimization for indexed nodes
The functions are made generic to accommodate both gates and basic
events. The current implementation only works for coherent trees. Final
pieces of the optimization are implemented, so this commit closes #100.

This optimization is sensitive to the initial structure of a fault
tree; however, currently no attempt is made to specifically manipulate
the fault tree before calling this optimization because it is uncertain
how the order of the other optimization algorithms affect each other for
certain kinds of fault trees. The application of the algorithms requires
careful consideration of the trade-offs and side effects.

Among the performance test cases, the 200-event fault tree is
the only one that improved substantially (at least 10x).