Make IGate change its logic for duplicate args
Duplicate arguments may reduce the number of arguments
and change the type of the gate if it has only one input.
The state of the gate can no longer be communicated through
the return value because it is not the only change possible.
The caller of the function must ensure
 that the final state of the gate is usable.