Make JoinGates() be aware of common gates
Some optimization and preprocessing algorithms
need common gates be preserved from coalescing.
JoinGates() is given a flag
to preserve common gates.

However, this also complicates the final result
of getting the layered structure of the graph.
To solve this complication,
Phase V is introduced,
which must cleanup after all the algorithms and phases.