Refactor all analyzers to use Settings
The settings are complex.
Instead of putting all of them in constructors for analyzers,
one Settings object for all the risk analysis
can be used for all sub-analyses.

Fault tree analyzer, probability calculator, and uncertainty analyzer
constructors are changed.

Redundant checks for correct settings are removed
because all the settings are validated by the Settings object.