Make DetectDistributivity consider more options
This commit makes DetectDistributivity a bit smarter.
The optimal solution for distributive setups is complicated
because it is a decision problem.
The current strategy emphasizes
the number of possible restructuring of the graph,
the number of gates involved in transformations,
the number of arguments that must be moved around.
There may be more important parameters
for distributivity solutions (maybe even for the single right solution?)
like the number of new clones of existing gates,
which may make the graph more complex for other preprocessing steps.

This commit also fixes a bug of not marking new clones,
which breaks the contract
and later leaves the subgraph untouched by other preprocessing steps.