Refactor the decision logic in MergeCommonArgs
The decision logic is separated into a function.
The strategy is similar to
the optimal factorization of common args in distributivity.
This fixes the originally flawed assumption
that if common args are subsets,
the common parents are supersets.

The current decision logic is dumb
and does not consider optimal options
with better modularity and arguments.
This is a complex problem
that will be tackled incrementally in future commits.