Improve the common node decomposition algorithm
The problem with the Shannon decomposition of common nodes
is that it introduces new clones of existing shared gates,
which makes the Boolean graph larger and more complex
for other preprocessing techniques.

The new algorithm avoids
the introduction of new clones as much as possible
by working on shared common gates only within the subgraph.
The exception is only made for shared parents of the common node.
These parents will get cloned
if they exist outside of the subgraph.

In addition, the repeated traversal and cloning is avoided
with memoisation and marking ancestor gates upon traversal.

Techniques of module detection are applied
to efficiently detect the residence of the shared gates
within the graph and subgraph.

OptiValue tricks of Boolean optimization is used
to efficiently detect ancestors and decomposition setups.

The performance hit for the analysis is around 10%.
However, the speed gain for preprocessing is exponential.
It is now possible to preprocess graphs with 1000s of nodes.
(A fault tree with 4000 gates and 2000 basic events runs in 10s)