Add gate substitution discovery in gate merging
This is an optimistic heuristic
that instead of merging common arguments,
substitute those common arguments with an existing gate
if possible.
It makes the common argument merging run 3-4 times faster,
but there is 10-20% performance hit in analysis.
The reason for the regression is
that other merging heuristics are not aware of substitutions
to produce optimal results.