Make arg merging heuristic favor the num parents
This is generalization of the heuristic
that favored the modules.
The new heuristic tries to find the optimal solution
by observing the reduction of the number of parents
of to-be-merged common arguments of a group of common parent gates.

The benefit of this heuristic is that common arg merging runs faster
and there is no penalty for analysis most of the time.