Use template base class for BDD NonTerminal Vertex
Template base class (CRTP) helps extend BDD
to other kinds of BDDs like Zero-Suppressed BDD.
The derived classes need only to put
the logic (role/meaning) of the non-terminal vertex;
whereas, the base class handles
the general operations of non-leaf nodes in BDD DAG.

The reason for the choice of templates is
to do a trick with smart pointers and
to avoid virtual functions.
Even though the template base class captures
the general patterns of BDD,
more advanced and specialized kinds of BDD
may want to avoid deriving from this base class.