Add BDD-to-ZBDD conversion procedures
The reduction rules of ZBDD are applied at once upon conversion.
It might also be possible to divide the conversion into two phases.
The first phase would be BDD-to-ZBDD cloning,
and the second phase would be ZBDD reduction.

It is not yet clear
how ZBDD and BDD will diverge on their applications for analysis.
The current conversion is good enough for testing.