Redesign the BDD classes with an RTTI hack
It is hard to separate Terminal and NonTerminal vertices
for algorithms at compile time.

The original approach with compile-time knowledge of types
leads to combinatorial explosion of overloads for all sorts of cases.
Even simple algorithms had to produce at least 4 overloads
to handle the combination of types of the arguments
while the actual work was within couple of lines.

Moreover, since it is impossible in C++ to overload on return types,
two distinct member functions return Terminal and NonTerminal children.
All callers of these functions had to be aware of the fact
that 'nullptr' means the callers need to use the other member function
so that not to lose the type information.
This choice has lead to
proliferation of 'nullptr' based logic in algorithms.

Avoiding RTTI seems to have been
a wrong decision for the design of BDD classes.
It is highly likely
that Terminal and NonTerminal classes are
all it is needed for algorithms,
and no new Vertex classes will be introduced
to break RTTI based code.

RTTI is used without dynamic casts or typeids.
Instead, a workaround with explicit 'type' field for Vertex is chosen.
The algorithms may choose to use static casts
to recover the type information.

This RTTI-based approach completely
avoids 'nullptr' based logic or functions.
NonTerminal class removes itself from handling the logic of algorithms.
Finally, this approach keeps intact
the original design goal of
clearly separating Terminal and NonTerminal Vertex class interfaces
without C-like NULL edges and three-valued-logic hacks.