Add fault tree preprocessing for BDD analysis
BDD based analyses may not need
the full normalization,
gate coalescing,
and complement propagation
as MOCUS does.

The current Preprocessor class is
not designed to be flexible
for generation of arbitrary combinations of preprocessing techniques.
The application of preprocessing techniques
will evolve together with the BDD based analysis.

This commit addresses issue #125.