Introduce the complement edge for BDD ITE vertex
The goals of application of a complement edge is
to allow constant time complementation
and to increase the sharing of existing nodes.

Complement edge aware Apply() and other procedures
may optimize algorithms by noticing equivalent operations,
such as OR(F, ~G) = ~(AND(~F, G)).

ZBDD is not yet considered with complement edges
because it is not clear
whether ZBDD needs it for monotonic (coherent) analysis.