Separate the Preprocessor for BDD into a new class
This separate class can introduce its own algorithms
to help BDD analyses run more efficiently.
It should not be a problem
if the logic of preprocessing gets coupled with the BDD data structure
because this specialized class is not expected to be used
outside of BDD-based applications.
This commit addresses issue #125.