Use BDD graph marking strategy to avoid ClearMark
This strategy makes
the users of BDD graph keep track of the current mark of the graph.
This decision is made to avoid double-traversal of the graph
as is done in Preprocessing.
The reason for this strategy is
that there are not many mark-dependent BDD graph functions
in comparison to Boolean graph functions.
In addition,
the mark-dependent BDD graph functions tend to be localized.

The side-effect of this strategy is
that the localized mark-dependent BDD graph functions
will be coupled by the state of the mark.

This strategy may be changed if coupling becomes an issue.