Make FTA class calculate MCS probability
This is a part of an effort (issue #117) to make FaultTreeAnalysis
aware of cut set probabilities and contributions.
ProbabilityAnalysis may be free from doing this simple work
and let FaultTreeAnalysis report its MCS probabilities.
However, the Reporter and RiskAnalysis classes are not ready
for this change.

Making FaultTreeAnalysis calculate quantitative values
may seem to be a design smell
and violation of separation of concerns
between ProbabilityAnalysis and FaultTreeAnalysis,
but the current approach is ugly as well,
and it is not yet clear
if the new approach is uglier.