Fix lambda function formatting with ClangFormat
The formatting of lambda functions is tricky.
The LLVM Coding Standards have
a pretty solution for this problem.
Depending on the place of definition of lambda functions
(inside std algorithms, last argument, non-last argument, etc.),
it is treated as a free function restricted by indentation.

Google C++ standard seems to use the same strategy,
but there is no explicit extensive coverage for particular setups.