Fix the Importance factor formulas
The derivations for factors had wrong assumptions,
such as coherence of the fault tree.
There were numerous mistakes in interpreting
conditional probabilities
and their relationships with the importance factors.

This serious problem was never caught because of
the lack of advanced benchmarking tests.
Unfortunately, there's still no tool to use to benchmark against
with exact probabilities and importance factors
for non-coherent fault trees.

Only very trivial cases are tested,
so there may well be more hidden problems
with the importance factor calculations.
However, this fix is good enough to close #127.