Calculate importance factors in one-pass from BDD
Instead of calculating conditional probabilities,
Importance analysis can find
the Marginal Importance Factor in one traversal over BDD.
This MIF, p(e), and p(S) can be used to derive the other factors.
This approach cuts the running time at least 2 times.

Calculation of importance factors in one-pass makes
the importance analysis more aware of and dependent on BDD.
If more optimization is desired
(especially with uncertainty analysis of importance factors),
Importance analysis is ready to use
Module visit information (from Preprocessor)
to cut the traversal of certain sub-graphs.
This optimization should give big performance boost;
however, it will effectively couple
Importance analysis with Preprocessor.
For this reason, this optimization approach is not in this commit.

This BDD algorithm is good to close #119.