Add Analysis call to BDD class
BDD can be asked to run the Qualitative analysis.
It uses facilities of ZBDD to run Quantitative analysis.

In order to access ZBDD,
BDD must store a pointer to ZBDD;
however, smart pointers can't be used due to circular inclusion problem
for ZBDD class uses BDD data structure class to create itself.
The right way to solve this problem
is to separate the BDD data structure from the Analysis.
For example, BDD based analysis can be placed in a different file,
or BDD data structure classes can be placed in a separate file.

The current decision is to keep the file structure unchanged
and to do the work with
forward declarations and manual memory management (one owning pointer).
This may not be the right and C++11 approach,
but this solution is the simplest.