Simplify *Analysis template interfaces
By removing generalized container for cut sets,
template specialization code is placed in the source file.
The cut set is fixed to be a vector of integers (indices).
This is a simple contract with FaultTreeAnalysis algorithms,
which also removes the need of Quantitative analyses
to be aware of the logic of FaultTreeAnalysis.