Cleanup FaultTreeAnalysis' CutSet responsibilities
Post-analysis is delegated to the reporter.
Any user of FaultTreeAnalysis can do post-analysis with helper tools.
However, FaultTreeAnalysis will not try to predict
what the interest of the caller is.
The only responsibility of the FaultTreeAnalysis is to get cut sets.
This also removes the design smell
of FaultTreeAnalysis calculating cut set probabilities.