Enable Boolean optimization for non-coherent graph
It is not yet fully tested or testable
because of the inability to compute prime implicants.
CEA9601 test is failing in both fault tree and quantitative analysis.
However, the number of cut sets are around 50k,
which is a good sign.
The probability value has the same magnitude but off by 50%.
There may also be floating point arithmetic problems
in computing probability values.