Handle default cases in switch statements
Compiler warnings used to be suppressed for this case.
However, it is a good practice to cover all switch cases.
In addition, GCSG mandates it for enumerated values.
Basic assertion should help with bugs
when new enumerated values are added to gate types.

Preprocessor, MOCUS, and Boolean graph codes are mostly cleaned up.