Fix ZBDD result node registration on construction
It is a bug in BDD to ZBDD conversion.
If the result node of an operation already in the graph,
the result is not registered,
but the node in the graph is returned.
This leads to re-computation of similar functions,
and the table of results ends up containing a null pointer.