Calculate proper MCS in MOCUS by default
Minimal cut sets include only positive literals.
For non-coherent fault trees,
this is a conservative approximation.
More exact approach will be achieved with prime implicants.

The current implementation is rough,
but it passes all tests to close #123.

Restriction of minimal cut sets to positive literals
solves many flaws in SCRAM as in #92.
The solution is good to close #92.