Fix a design flaw in Decomposition preprocessing
The fundamental flaw in Decomposition preprocessing step
is an assumption that
if a source node is within graph's visit bounds,
the failure propagation will not affect anything outside of the graph,
so changing the node is safe.
However, not all the nodes in the path to the source node
are tested for being in the same graph,
so if any of ancestor nodes of the source node
is shared with another graph,
the result of preprocessing is not semantics preserving.

The solution to this problem
is to conservatively clone the source without any assumptions.
This is 10-20x performance hit for the Decomposition preprocessor,
but correctness is more important.