Generalize the interface of Boolean optimization
The Boolean optimization can be generalized
for propagation of success.
The documentation is updated together with the interface
to give the theoretical reasoning for the generalization.

The interface is generalized by noting
that a particular State is propagated,
not necessarily Failure.

Propagation of success is not, however, applied in implementation.
Moreover, it is not yet clear
if the propagation of success should be applied at all.