Adjust Boolean optimization for non-coherent cases
This is an experimental feature
to assess the effect of Boolean optimization failure propagation
on non-coherent models.

The current implementation worsens the performance of coherent cases
because of the need to generalize the technique.

There's no observed improvement for non-coherent models as well.
It is currently hard to test without having fast analysis.