Simplify Error class' code
The error classes are simplified with constructor inheritance.
Since all current error classes follow the same pattern,
it would be easy to generalize them with a macro.
However, the problem with this option
is that Doxygen is not able to expand macros
that generate the error classes,
so these error classes will not show up in API documentation,
which is a huge price to pay for simplifying some extra code.