Fix ZBDD Subsume operations
The defect is found in the implementation of the algorithm.
The problems are in the case
with the subsume of nodes with the same indices.
Extra subsume operation was missing due to faulty assumptions.

Reduction rules and assumptions are tested with a new test function.
This function helps find defects in the structure of ZBDD.