Enable tests for analysis of CEA9601 with BDD
The fix of the application of the limit order for non-coherent graphs
made possible the proper calculation of CEA9601 for the first time
in SCRAM's existence.

The benchmark and performance tests are very helpful
for verification of BDD for FTA (#121).
Performance logs are updated to track the progress and regress.