Make BDD garbage collector work on compute tables
The implementation is too eager to clean
and inefficient with tracking the tables to clean.
The memory use for the most intensive benchmarks
has actually increased by around 20%.
Speed for the most intensive benchmark has decreased by 50%.

Even though the performance results are very bad,
there is a lot of room for improvements.

The effect of the garbage collector is logged in the speedlog.
There is clear improvement in the number of items in BDD hash tables.
This garbage collector is an important part of issue #121.