Make the ZBDD Apply procedure minimize its results
The minimization works with non-coherent sets.
The memory utilization improvement for ZBDD creation from a PDAG
is around 7x.
The speed improvement is around 3x.
The number of nodes created for ZBDD has dropped by a factor of 10.

The downside of the current approach
is a performance hit for MOCUS using ZBDD.
The ZBDD use in MOCUS must change in order to avoid this hit.

Issue #126.