Add a script to generate a fault tree of N Queens
This is a fault tree representation of the N Queens problem.
The problem can serve as a benchmark for BDD implementation.
MOCUS and ZBDD also solve the default (8 Queens) problem
in a short time as BDD (< 0.2s).

The major surprise came from the Preprocessor.
It takes from 0.5s to 1s depending on the analysis algorithm.