Refactor Gate-to-BDD/ZBDD memoisation strategy
The use count of the gate as an argument is taken into account.
If the gate is known not to be referenced in future,
the memoised results are discarded,
which triggers potential garbage collection of BDD/ZBDD vertices.
The peak memory consumption dropped by around 10%.

The downside of this strategy is
that garbage collected vertices are likely to be created again later
for another calculation because future gates share common structure.
As a result, 1-5% more vertices are created per computation on average.