Refactor ZBDD complement and module elimination
The elimination process does not proceed into sub-modules.
The caller must make sure
that those sub-modules have been already processed.
This refactoring helps share (reuse)
the same approach with the CutSetContainer in MOCUS.

The big problem with these complement and constant module elimination
is the inefficiency and strictness
(wasteful operation is the branch doesn't need any work).
To fix the problem,
the elimination strategy must take advantage of
the DAG nature of BDDs.

Issue #124