Add prime implicant calculation for simple cases
Many facilities are designed with minimal cut sets in mind,
so they are not ready to handle complex cases yet.
For example, complemented modules are currently ignored.
This can be considered a bug
even for minimal cut set calculations with BDD.

The initial ad hoc design for prime implicants is ugly.
BDD and ZBDD both need to cooperate to calculate prime implicants,
so ZBDD loses strict const guarantees for BDDs.
The function overloads for prime implicant calculations in ZBDD
are named in a bug-prone way:
functions for prime implicants end with 'PI'.

Initial set of simple tests are enable for prime implicants.

Issue #122.