Generalize cut sets to Boolean products
"Boolean products" are used for generality in analysis facilities.
Only in cases where facilities specifically deal with "cut sets",
"cut sets" are used in API and documentation.

The generalization is a part of an effort
to clarify the distinction between cut sets and implicants
in the documentation and API (Issue #122).