Implement lazy modules with adjusted cutoffs
The feature is implemented for qualitative analysis with BDD.
ZBDD class takes care of the requested behavior.
Unfortunately, this performance optimization is hard to test
due to the need for white box testing.
Currently, only verification inputs are provided for testing.
The logging information can be used to verify the lazy behavior.

MOCUS and ZBDD based analyses do not have the implementation
due to currently unfixed (messy) logging of nested analysis,
which makes verification of the behavior hard (unintelligible).