Refactor GetExpression functions in Initializer
Instead of branching with many if-else statements
to construct various types of expressions,
a map of function objects is provided to do the construction.

This approach is rather "too clever" for this simple task
due to the use of partial template specializations,
C++11 variadic templates, static container with dynamic allocation.
However, with dozens of OpenPSA MEF expression types to be introduced,
this approach with templates and maps keeps the code manageable.