Fix Combination vs. Voting vs. Atleast vs. K/N
There's no Boolean operator for this K/N gate.
In order to keep the SCRAM code consistent,
the following guidelines are added.

In internal code and API,
"vote" is required.
In code facing the OpenPSA MEF,
"atleast" may be used.

The code and documentation are fixed with the new guideline.