Refactor Reporter to use XmlStreamElement
XmlStreamElement replaces the XML DOM elements
for reporting.
It is almost drop-in replacement for reporting with streams.

The XmlStreamElement is refactored to be safer to use.
It prevents concurrent writing
with two child elements on the same scope.

The speed of reporting improved to around 200,000 cut sets per second,
and memory overhead dropped to almost nothing
because of the use of streams.