Add support for complement args in the shorthand
Arguably, this is the only useful case for nested formulae
in the shorthand format.
The use of complement arguments in formulas
requires a change in formatting of NOT and NULL gates
to avoid the ambiguity of NULL(NOT arg).

The feature is reintroduced
to support the logged print of BooleanGraph of the core code,
which inherently works with complement arguments.