Hide Vertex ptr typedefs in BDD and ZBDD
The aliases are hidden but public in Bdd and Zbdd classes.
This approach forces the code to be clear
about which vertex it is using.

The current implementation suffers from ambiguity
because of the Bdd and Zbdd vertex class hierarchy.
Many interfaces work with base class instances;
however, the logic heavily relies
on the assumptions about the derived class instances.
Therefore, the interfaces misrepresent the actual functionality.
The functions taking the base class are not general enough
to take any class deriving from the base class.