Implement custom BDD/ZBDD UniqueTable hash table
The implementation replaces the standard unordered map
as the base hash table for UniqueTable.
The unique table is used to keep BDD/ZBDD reduced,
and this logic is integrated into the hash table
to keep it smaller and faster.

The custom implementation gives a full control
over the UniqueTable performance.
It will be easier to take advantage of BDD-specific use cases
and improve the performance of the table.
Moreover, the hash table characteristics
can be exactly tailored for BDD-specific use patterns.

The implemented UniqueTable is intrusive.
It relies on BDD vertices to get keys in the table.
There is no key-value pairs.
This approach can be considered
an integration of the BDD graph and the UniqueTable.

The speed improvement for large problems ranges from 5 to 20%.
The memory utilization improved (for large problems) by ~20%.
The performance effect is not as noticeable for smaller problem sets
because the hash table operations are not as big contributors.

CPU cache-misses are recorded in the performance logs
with the "perf" tool.
This metric is useful to see
the effect of working with a graph (linked data structure) and table.