Make Bdd/Zbdd UniqueTable growth factor dynamic
The growth factor is adjusted
depending on the current size of the graph.
This is an attempt to take into account
the growth patterns of BDDs.

The hash table resizing and rehashing tends to dominate
the performance of small to medium graphs in its initial phase.
The dynamic growth factor attempts to compensate for the growing BDD
by gradually decreasing the growth factor.

This is not a hard-rule.
Some BDDs may behave erratically instead of growing gradually.
The current approach is just an heuristic
that has evident speed benefits for large graphs (10 - 20 %).