Make ZBDD-set-order cut-off application strict
The strict cut-off application makes ZBDD operations more efficient
because of the fewer sets to operate upon.
However, since the set order calculation is conservative,
there are sets with higher order
that are revealed only at the end of the analysis.
These sets are treated conservatively in the strict mode as well.

The ZBDD-related performance improvement is around 20%.
The performance of ZBDD has become good enough
to add more benchmarking and performance tests
that run in reasonable time.