Implement vector-based LinearMap
The vector-based map can be used
when the key-value pairs are small and simple,
when the number of entries is small,
and when the order of insertion matters.
When these constraints are satisfied,
the LinearMap performs better than or as well as
std::map, std::unordered_map, and boost::flat_map.