Fix DecompositionProcessor side-effect prevention
The decomposition processor tries to avoid side-effects
outside of the sub-graph.
The side-effect prevention check helps avoid extra cloning,
and the decomposition success depends on the proper cloning.

The logic of the side-effect prevention is faulty
by assuming that sub-graph root visit times
are enough to determine
if the no-clone opportunity can be exercised.
It turns out to require
the check of the whole hierarchy of ancestors.

The current fix is not efficient;
it has exponential time complexity.
However, this correct implementation
should lead to better algorithms with better performance.

This fix closes #128.