Fix performance regression of Model::GetEntity()
C++11 lambdas return by copy.
Using lambdas as wrappers to access member functions
is a very bad idea for this reason.
The lambda cannot automatically get
the const reference qualified return type,
instead it just copies the return,
which led to awful performance regression
in the getter functions of Model.

The lambda wrapper is stripped away to regain the performance.
Plain pointers to member functions are used instead.
This approach is not convenient with an ugly syntax in C++,
but it gets the most efficiency.