Implement CacheTable for BDD
CacheTable is a hash table without collision resolution.
Upon collision, old values are removed from the table.
This table is used to memoize the BDD Apply() results,
which means that the complexity guarantees do not hold anymore.
That is, the worst case complexity of BDD::Apply() is exponential.
However, it is very unlikely in practice.

The CacheTable API is made exactly as the STL hash tables
for two reasons.
First, should the Bdd::Apply() performance become increasing intolerable,
it would be easy to swap the CacheTable with a normal hash table.
Second, upon performance benchmarking,
the STL hash tables are used as a baseline.

The memory use dropped by 10%,
and speed increased by 20% for demanding BDD analyses.

Surprisingly, the speed effect is not what was expected.
The speed has been expected to degrade
due to less effective memoization.
This result is likely a tell
that the current BDD implementation
has not been using memoization effectively in first place.