Remove assertion from MOCUS sort order predicate
The predicate is passed to std::sort.
On AppleClang,
the sort is comparing the pivot with itself,
which is inefficiency,
or it is doing something different from quick sort.
My assertion was that the arguments passed to the sort predicate
are different objects. (which is always the case with GCC).
Probably, this is not required by the C++ standard,
so it is not an STL bug on AppleClang.