Update performance metrics on Ubuntu 16.04
The performance got worse on Ubuntu 16.04 by around 10%.
The very likely reason is ordering change of BDD and ZBDD vertices.
The ordering is affected by the iteration of standard unordered maps,
which is implementation dependent.

Around 20% IO improvement is observed on Ubuntu 16.04,
which is not much relevant to the analysis.