Reformat input files expecting Transfer symbols
These input files are from OpenFTA
that doesn't have a concept of PDAG,
so the shared gates are separated into separate fault trees.

These inputs are reformatted to have fault trees or components
with the root gate as the exported (transfer-out).
The symbol is ultimately graphical.
However, the current implementation of Grapher abuses this symbol
for any shared gates as in OpenFTA.