Separate Cycle API from Events and Expressions
The cycle detection
is not the responsibility of the Event or Expression.
The Initializer employs other facilities to detect cycles
without adding API to the cycle-prone classes.
The Cycle header facilities just overload
their detection, retrieval functions.

It is not expected that more cyclic graphs will be added in future;
therefore, this may be a lucky case to accept the expression problem.