Implement space-efficient Product for Literals
The results of analysis are stored in Products for reporting.
The most efficient approach would be to utilize ZBDD directly;
however, the current set of sets approach is easy on other parts
of the reporting facilities.

This implementation of Product container cuts memory use by 40-50%.
However, it is very restrictive with constraints.
The analysis is limited to 32 literals only.
It is in fact uncommon to reach this limit,
so the constraint does not break any current analysis.

The speed penalty of the space optimization is negligible,
considering the efficient handling of a pair of value types.