Use LinearMap in PDAG Gate arg containers
The linear map for gate arguments is more stable than the hash table.
Considering that the number of arguments is low,
the speed and space improvements are up to 2x for the use cases.

The preprocessing technique can be more stable with the linear map
instead of the hash table across platforms and various implementations.
Because of the switch to the linear map,
the preprocessed fault tree before analysis has changed heavily.
The variable ordering of BDD/ZBDD has been influenced heavily
by the change of the data structure for gate argument management.

The performance impact of the change is varying
from -10% to +10% in speed and memory for overall analysis.