Introduce specialized Element and Id Tables
This tables specialize in keeping elements unique
by their names or ids, respectively.
Instead of having hash tables with a repeated information,
such as the element name or id, as a key,
boost::multi_index is employed to index elements
by their properties.

This approach should make element database management more abstract
(similar to relational databases)
instead of messing with key-value pairs.

Applications of these abstract tables should result in simpler code,
but speed gains are not expected.
It is possible to see some performance regression
due to one extra pointer chasing (cache misses)
in order to get keys for tables from the element fields,
which are stored as pointers in the database.