Fix naming of lambda expression variables
The rule to name lambda expression variables
as functions with the mixed case style has been abandoned
in favor of the snake case style
for consistency with the rest of local variables.
However, the name is a verb instead of a noun,
which acts like an indicator for a cheap function.
Since the result of non-local functions follow the mixed case style,
the name format of the lambda expression variable stands out,
which makes a good hint/pattern for a reader.