Change the style conventions for qualified calls
The current practice of qualifying all function calls
except for global and ADL functions
has made code noisy.

In order to reduce the noise
and still keep the code readable
without hopping through included headers,
the rules are swapped for calls of member functions and globals.
This scales towards solving the most common case,
which calling another member function within a member function.
Calls to global functions need to be qualified with namespaces.

The rules for the ADL calls, virtual functions, and inherited functions
did not change.

New rules leverage the context of calls
by requiring the reader
to be aware of the function being defined in the code.